Make basic store via introspected
