+++ package.json
@@ -69,20 +69,20 @@
   "homepage": "https://github.com/dy/spect#readme",
   "type": "module",
   "dependencies": {
-    "delegate-it": "^1.1.0",
+    "delegated-events": "^1.1.1",
     "fast-deep-equal": "^2.0.1",
     "fast-on-load": "^1.2.0",
     "htm": "^2.2.1",
     "immutable-tuple": "^0.4.10",
     "incremental-dom": "^0.6.0",
     "insert-styles": "^1.2.1",
+    "introspected": "^0.2.6",
     "kebab-case": "^1.0.0",
     "nanomorph": "^5.4.0",
     "negative-index": "^1.0.3",
     "preact": "^10.0.0-rc.3",
-    "react": "^16.9.0",
-    "react-hookstore": "^1.4.2",
-    "scope-css": "^1.2.1"
+    "scope-css": "^1.2.1",
+    "selector-observer": "^2.1.6"
   },
   "devDependencies": {
     "@pika/pack": "^0.5.0",
