+++ src/$.js
@@ -59,6 +59,16 @@ const _ = {
   }
 }
 
+// select elements
+export function $(sel) {
+  return document.querySelector(sel)
+}
+export function $$(sel) {
+  return document.querySelectorAll(sel)
+}
+
+
+
 export default _.$
 
 class $ extends Array {
