+++ src/core.js
@@ -1,7 +1,4 @@
 import equal from 'fast-deep-equal'
-import { render as prender, h as hs } from 'preact'
-import htm from 'htm'
-import tuple from 'immutable-tuple'
 import { isPrimitive } from './util'
 
 const isStacktraceAvailable = !!(new Error).stack
@@ -50,7 +47,7 @@ function callsiteId() {
   return key
 }
 
-function checkDeps(deps, destroy) {
+export function deps(deps, destroy) {
   if (!current) return true
 
   let key = callsiteId()
@@ -108,73 +105,4 @@ function queue(fn) {
       planned = null
     })
   }
-}
-
-// render vdom into element
-export function render(vdom, el) {
-  if (typeof el === 'string') el = $$(el)
-  else if (el instanceof Node) prender(vdom, el)
-  else el.forEach(el => prender(vdom, el))
-}
-
-
-// build vdom
-export const html = htm.bind(h)
-
-// select elements
-export function $(sel) {
-  return document.querySelector(sel)
-}
-export function $$(sel) {
-  return document.querySelectorAll(sel)
-}
-
-function h(tagName, props, ...children) {
-  if (typeof tagName !== 'string') return hs(...arguments)
-
-  if (!props) props = {}
-  let [tag, id, classes] = parseTagComponents(tagName)
-  if (!props.id) props.id = id
-  if (!props.class) props.class = classes.join(' ')
-
-  return hs(tag, props, ...children)
-}
-
-function parseTagComponents(str) {
-  let tag, id, classes
-  let [beforeId, afterId = ''] = str.split('#')
-  let beforeClx = beforeId.split('.')
-  tag = beforeClx.shift()
-  let afterClx = afterId.split('.')
-  id = afterClx.shift()
-  classes = [...beforeClx, ...afterClx]
-  return [tag, id, classes]
-}
-
-
-// turn function into a web-component
-const _destroy = Symbol('destroy')
-export function component(fn) {
-  class HTMLCustomComponent extends HTMLElement {
-    constructor() {
-      super()
-    }
-    connectedCallback() {
-      // take over attrs as props
-      [...this.attributes].forEach(({ name, value }) => {
-        this[name] = value
-      })
-      new Promise((ok) => {
-        setTimeout(() => {
-          ok()
-          this[_destroy] = fn.call(this, this)
-        })
-      })
-    }
-    disconnectedCallback() {
-      this[_destroy] && this[_destroy]
-    }
-  }
-
-  return HTMLCustomComponent
 }
